Set timezone to GMT as a reference
The time zone is usually detected as the current time zone of the computer. This can lead to some kind of trouble with daylight saving time. We set GMT because it is the only one that seems to be universal and well dealt with in R. See the doc for as.POSIXct:
      tz: A timezone specification to be used for the conversion, _if
          one is required_.  System-specific (see time zones), but '""'
          is the current timezone, and '"GMT"' is UTC (Universal Time,
          Coordinated).