Allow the camera to be above or below the arena
The various corrections need to be adapted to whether the compass is seen from underneath or above: this inverts the East and the positions have to be adapted to that.

In addition, when the compass is manual, we need to ask the user because we don't have information to guess that.