Allow to use bb without a compass track
This is a first dirty workaround for the situations where bb is used to analyse pictures taken in a non-rotating environment (e.g. a lab).

Allow to not have any compass track (convert error to warnings in bb).

When there is no compass track, just provide a fake 0 heading all the time.